requirements.txt
fastapi[standard]
requests
python-dotenv==1.0.1
Jinja2==3.1.4
